+++ VersionHistory.md
@@ -7,7 +7,8 @@ Describe changes here when they're committed to the `master` branch. Move them t
 Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
 * [patch] Add `params` support to method signature.
-* [patch] Add `Caller` attributes to method signature.
+* [minor] Add `Caller` attributes to method signature.
+* [minor] Support hyperlinks to URLs.
 
 ## Released